feat: consolidate button components into single directory
- Merge ui/button and ui/buttons into unified ui/buttons directory
- Move shadcn Button component and buttonVariants to ui/buttons
- Create centralized exports in buttons/index.js
- Update all imports to use destructured imports from @/components/ui/buttons
- Remove duplicate button directory structure

Closes button organization issue from refactoring plan
