+++ src/components/app/HelloWorld.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import BaseButton from "./buttons/BaseButton.vue";
+import { BaseButton } from "@/components/ui/buttons";
 </script>
 
 <template>
