+++ src/components/ui/buttons/index.js
@@ -1,6 +1,10 @@
 import { cva } from "class-variance-authority";
 
 export { default as Button } from "./Button.vue";
+export { default as BaseButton } from "./BaseButton.vue";
+export { default as ButtonText } from "./ButtonText.vue";
+export { default as NavButtonContext } from "./NavButtonContext.vue";
+
 
 export const buttonVariants = cva(
   "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0",
