+++ src/components/ui/calendar/CalendarNextButton.vue
@@ -3,7 +3,7 @@ import { reactiveOmit } from "@vueuse/core";
 import { ChevronRight } from "lucide-vue-next";
 import { CalendarNext, useForwardProps } from "reka-ui";
 import { cn } from "@/lib/utils";
-import { buttonVariants } from '@/components/ui/button';
+import { buttonVariants } from '@/components/ui/buttons';
 
 const props = defineProps({
   nextPage: { type: Function, required: false },
