+++ src/components/ui/datepicker/DatePickerComponent.vue
@@ -9,7 +9,7 @@ import { CalendarIcon } from "lucide-vue-next"
 
 import { watch, ref } from "vue"
 import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/buttons"
 import { Calendar } from "@/components/ui/calendar"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
